[spec] pdfrum-edit: the root `pub use` block is the surface, and it is the only one
§A.11's row: "`pub mod content`/`encrypt`/`font`/`import`/`write` → curated;
today each is `pub mod` *and* selectively re-exported, so the same items are
reachable two ways." They are now reachable one way. Every module is private
(`pub(crate)` where a sibling reaches across — `write` names `encrypt`,
`content::marks` names `write::object`, `import::nup` names `content::num`),
and the twenty-eight-name `pub use` block in `lib.rs` is the whole API.

**374 items and 16 `pub mod` become 116 and 1** — the largest proportional
reduction of the pass so far, and none of it a rename: every item the block
names keeps its name, its signature and its path, and the 258 that went are
items no caller outside the crate ever reached. `crates/pdfrum/tests/reexports.rs`
needed no edit and the facade needed no edit, which is the measurement that
proves the second path was unused rather than merely unfashionable.

§A.11 said 5 `pub mod`; measured it is 16, because `content` and `write`
publish their own submodules too. The estimate counted `lib.rs` only.

**The bool→enum pass, which is what §A.11 asked for.** Its illustration is
`paint_operator(fill: FillRule, stroke: bool)` — "one argument is properly an
enum, its neighbour is a bool." Both arguments now read the same way: a
crate-private `Stroked { Yes, No }` with a `Stroked::of` narrowing at the one
call that has a `bool` to narrow, because `PathObject::stroke` lives in
`pdfrum-page` and belongs to step 9. Applying the same test to every other
`bool` in the surface, the full list and the ruling on each:

- **`clip_rule(even_odd: bool) -> FillRule`** — the crate's second
  bool-as-enum-argument, `content/emit.rs:422`, whose own doc said "spelled
  for a test to read". **No test read it, and nothing else did either.**
  Deleted, with `emit_page_objects` (a loop nothing calls) and `fresh_state`
  (`GraphicsState::default()` behind a wrapper). Three dead public functions
  that existed only because their module was public — the same thing §WP2
  found when `annot_dump` moved.
- **`SaveOptions::{keep_original, remove_security, subset_new_fonts}`,
  `ImportOptions::viewer_preferences`** — settings on an options struct, each
  a genuine independent yes/no with no third state hiding in it. They stay.
- **`FileId::rekeyed`, `Encryptor::encrypts_metadata`, `EditDoc::is_edited` /
  `is_removed`, `GidMap::is_empty`, `PageRange::is_empty`** — predicates and
  reported facts, not modes. They stay; §C.3 Tier 3 is about a `bool` that
  *discards a reason*, and none of these is a mutation result.
- Three `bool` parameters survive on private functions (`partition`'s
  `incremental`, `emit_colour`'s `stroking`, `named`'s `writing`). §A.11's
  rule is about the public surface and they are no longer on it; they are
  recorded here so the next reader knows they were looked at.

**The last `i32` goes.** `apply_rewrite` returned `BTreeMap<i32, i32>` and
`ContentsShape::with_removed` returned one — a map between `/Contents`
indices, signed only by contagion from `NO_CONTENT_STREAM`, which §C.3 item 1
deleted in step 9. Both are `BTreeMap<usize, usize>`, which retires an
`i32::try_from` guard that silently dropped an entry it could not narrow.
§C.3 item 16's "listed so the fix is not left half-done", finished.
`regen.rs`'s `pub stream` was already `Option<usize>`, as the brief said.

**What privatising revealed, and why ten files carry a `dead_code` allowance
rather than a deletion.** `dead_code` sees the library compiled without
`cfg(test)`, so an item only the crate's own unit tests exercise now reads as
dead. Twenty-three such items surfaced across ten files, in two kinds:

- **Test-only completions** — `ObjectMap::{len,is_empty}`,
  `ObjectOffsets::{numbers,len,is_empty}`, `ReachCounts::{reachable,
  multiply_referenced,is_shared}`, `trailer::{from_scratch,is_suppressed}`,
  `NupGrid::sheet_count`, `copy_inheritable`, `array_is_flat`. Each is
  pinned by a test beside it; each was reachable only because its module was
  public.
- **A save-path stage that is not wired.** `font/`'s `subset_tag`,
  `subset_name`, `strip_subset_prefix`, `is_opentype_cff`, `to_unicode_cmap`
  and `widths_array` implement SPEC §11 E4's re-keying contract — /W,
  /ToUnicode and the six-letter tag, re-keyed through the subsetter's GID map.
  They are the stage `SaveOptions::subset_new_fonts` names, and **nothing
  anywhere reads that field**: `grep` finds two hits, its declaration and its
  default. A public option that does nothing is a separate finding and not
  this step's to fix; deleting its machinery would be the wrong half to
  resolve it by.

Each allowance names its reason at the file it sits on.

+++ crates/pdfrum-edit/src/content/apply.rs
@@ -49,7 +49,7 @@ pub fn apply_rewrite(
     page_dict: &Dict,
     rewrite: &PageRewrite,
     shared: &ShareCounts,
-) -> BTreeMap<i32, i32> {
+) -> BTreeMap<usize, usize> {
     let mut shape = ContentsShape::read(page_dict, edit);
     let mut removed: BTreeSet<usize> = BTreeSet::new();
     let mut contents_changed = false;
